# pims/pims/openapi/api-specification.yaml
openapi: 3.0.3
info:
  title: Cytomine Python Image Managament Server PIMS
  description: |
    Cytomine Python Image Managament Server (PIMS) HTTP API. While this API is intended to be internal, a lot of the
    following specification can be ported to the external Cytomine API.

    # Additional API formats
    In addition to OpenAPI types and formats, the following table contains some additional formats used in this
    specification.

    | Type | Format | Examples | Specification |
    |------|--------|----------|---------------|
    | `string` | `range` | **"1:3"**, **"2:"**, **":"**| Inspired from Python slicing |
    | `string` | `wkt` | **"POINT (10 20)"** | [Open Geospacial Consortium](https://www.ogc.org/standards/sfa)|
    | `string` | `color` | **"#FF00AA"**, **"rgb(100,0,0)"**, **"red"** | [CSS Color Module Level 3](https://drafts.csswg.org/css-color-3/#colorunits)|
    | `string` | `colormap-id` | **"jet"**, **"!jet"**||
    | `string` | `filter-id` | **"yen"** ||
  version: 0.1.0
#servers:
#  - url: 'https'
...
paths:
...
  /image/{filepath}/annotation/mask:
    post:
      operationId: pims.api.annotation.show_mask
      tags:
        - Annotations
      summary: Get an annotation mask
      description: |
        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
        with body content must be used.**

        The mask is a generated image where geometries are filled by their respective `fill_color`. The background is
        black.

        The input spatial region is given by the rectangular envelope of all geometries multiplied by an optional
        context factor. The target size is given by one of the scaling factors (`size`, `width`, `height`, `zoom` or
        `level`).

        By default, a binary mask with white foreground is returned as the default fill color is white for every
        annotation.

        Annotation `stroke_width` and `stroke_color` are ignored.
      parameters:
        - $ref: '#/components/parameters/path-filepath'
        - $ref: '#/components/parameters/header-etag'
        - $ref: '#/components/parameters/header-x-image-size-safety'
        - $ref: '#/components/parameters/header-x-annotation-origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/mask-by-length'
                - $ref: '#/components/schemas/mask-by-width'
                - $ref: '#/components/schemas/mask-by-height'
                - $ref: '#/components/schemas/mask-by-zoom'
                - $ref: '#/components/schemas/mask-by-level'
      responses:
        200:
          $ref: '#/components/responses/annotation-crop'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
  /image/{filepath}/annotation/crop:
    post:
      operationId: pims.api.annotation.show_crop
      tags:
        - Annotations
      summary: Get an annotation crop
      description: |
        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
        with body content must be used.**

        The crop is similar to an image window but where the transparency of the background can be adjusted.

        The input spatial region is given by the rectangular envelope of all geometries multiplied by an optional
        context factor. The target size is given by one of the scaling factors (`size`, `width`, `height`, `zoom` or
        `level`).

        By default, the background transparency is set to 100 which is also known as *alpha mask*. When the
        background transparency is set to 0, foreground and background cannot be distinguished.

        Annotation `fill_color`, `stroke_width` and `stroke_color` are ignored.
      parameters:
        - $ref: '#/components/parameters/path-filepath'
        - $ref: '#/components/parameters/header-etag'
        - $ref: '#/components/parameters/header-x-image-size-safety'
        - $ref: '#/components/parameters/header-x-annotation-origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/crop-by-length'
                - $ref: '#/components/schemas/crop-by-width'
                - $ref: '#/components/schemas/crop-by-height'
                - $ref: '#/components/schemas/crop-by-zoom'
                - $ref: '#/components/schemas/crop-by-level'
      responses:
        200:
          $ref: '#/components/responses/annotation-crop'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
        406:
          $ref: '#/components/responses/problem-not-acceptable'
  /image/{filepath}/annotation/drawing:
    post:
      operationId: pims.api.annotation.show_drawing
      tags:
        - Annotations
      summary: Get an annotation drawing
      description: |
        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
        with body content must be used.**

        Get an annotation crop (with apparent background) where annotations are drawn according to their respective
        `fill_color`, `stroke_width` and `stroke_color`.

        The input spatial region is given by the rectangular envelope of all geometries multiplied by an optional
        context factor. The target size is given by one of the scaling factors (`size`, `width`, `height`, `zoom` or
        `level`).
      parameters:
        - $ref: '#/components/parameters/path-filepath'
        - $ref: '#/components/parameters/header-etag'
        - $ref: '#/components/parameters/header-x-image-size-safety'
        - $ref: '#/components/parameters/header-x-annotation-origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/drawing-by-length'
                - $ref: '#/components/schemas/drawing-by-width'
                - $ref: '#/components/schemas/drawing-by-height'
                - $ref: '#/components/schemas/drawing-by-zoom'
                - $ref: '#/components/schemas/drawing-by-level'
      responses:
        200:
          $ref: '#/components/responses/annotation-crop'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
#  /image/{filepath}/annotation/spectra:
#    post:
#      operationId: pims.api.annotation.show_spectra
#      summary: Get annotation spectra
#      description: |
#        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
#        with body content must be used.**
#
#        Get spectrum of every pixels belonging to the annotation(s).
#      tags:
#        - Annotations
#      parameters:
#        - $ref: '#/components/parameters/path-filepath'
#        - $ref: '#/components/parameters/header-etag'
#      requestBody:
#        required: true
#        content:
#          application/json:
#            schema:
#              oneOf:
#                - title: spectra-by-length
#                  allOf:
#                    - $ref: '#/components/schemas/annotations-object-basis'
#                    - $ref: '#/components/schemas/target-length-object-required'
#                    - $ref: '#/components/schemas/spectral-window-plane-selection'
#                - title: spectra-by-width
#                  allOf:
#                    - $ref: '#/components/schemas/annotations-object-basis'
#                    - $ref: '#/components/schemas/target-width-object-required'
#                    - $ref: '#/components/schemas/spectral-window-plane-selection'
#                - title: spectra-by-height
#                  allOf:
#                    - $ref: '#/components/schemas/annotations-object-basis'
#                    - $ref: '#/components/schemas/target-height-object-required'
#                    - $ref: '#/components/schemas/spectral-window-plane-selection'
#                - title: spectra-by-zoom
#                  allOf:
#                    - $ref: '#/components/schemas/annotations-object-basis'
#                    - $ref: '#/components/schemas/target-zoom-object-required'
#                    - $ref: '#/components/schemas/spectral-window-plane-selection'
#                - title: spectra-by-level
#                  allOf:
#                    - $ref: '#/components/schemas/annotations-object-basis'
#                    - $ref: '#/components/schemas/target-level-object-required'
#                    - $ref: '#/components/schemas/spectral-window-plane-selection'
#      responses:
#        200:
#          $ref: '#/components/responses/hyperspectral-window'
#        304:
#          $ref: '#/components/responses/not-modified'
#        400:
#          $ref: '#/components/responses/problem-bad-request'
#        404:
#          $ref: '#/components/responses/problem-not-found'
#  /image/{filepath}/annotation/footprint:
#    post:
#      operationId: pims.api.annotation.show_footprint
#      summary: Get annotation spectral footprint
#      description: |
#        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
#        with body content must be used.**
#
#        Get annotation spectral footprint. **At least one dimension must have a reduction
#        operation.**
#
#        The input spatial region is given by the rectangular envelope of all geometries.
#      tags:
#        - Annotations
#      parameters:
#        - $ref: '#/components/parameters/path-filepath'
#        - $ref: '#/components/parameters/header-etag'
#      requestBody:
#        required: true
#        content:
#          application/json:
#            schema:
#              oneOf:
#                - title: footprint-by-length
#                  allOf:
#                    - $ref: '#/components/schemas/annotations-object-basis'
#                    - $ref: '#/components/schemas/target-length-object-required'
#                    - $ref: '#/components/schemas/footprint-plane-selection'
#                - title: footprint-by-width
#                  allOf:
#                    - $ref: '#/components/schemas/annotations-object-basis'
#                    - $ref: '#/components/schemas/target-width-object-required'
#                    - $ref: '#/components/schemas/footprint-plane-selection'
#                - title: footprint-by-height
#                  allOf:
#                    - $ref: '#/components/schemas/annotations-object-basis'
#                    - $ref: '#/components/schemas/target-height-object-required'
#                    - $ref: '#/components/schemas/footprint-plane-selection'
#                - title: footprint-by-zoom
#                  allOf:
#                    - $ref: '#/components/schemas/annotations-object-basis'
#                    - $ref: '#/components/schemas/target-zoom-object-required'
#                    - $ref: '#/components/schemas/footprint-plane-selection'
#                - title: footprint-by-level
#                  allOf:
#                    - $ref: '#/components/schemas/annotations-object-basis'
#                    - $ref: '#/components/schemas/target-level-object-required'
#                    - $ref: '#/components/schemas/footprint-plane-selection'
#      responses:
#        200:
#          $ref: '#/components/responses/hyperspectral-footprint'
#        304:
#          $ref: '#/components/responses/not-modified'
#        400:
#          $ref: '#/components/responses/problem-bad-request'
#        404:
#          $ref: '#/components/responses/problem-not-found'
#  /image/{filepath}/spectral/footprint:
#    post:
#      operationId: pims.api.spectral.show_spectral_footprint
#      summary: Get spectral footprint
#      description: |
#        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
#        with body content must be used.**
#
#        Get spectral footprint. **At least one dimension must have a reduction
#        operation.**
#
#        ***Example 1:***
#        * X, Y: range of pixel coordinates with `AVG` reduction
#        * C: range of channels
#        * Z: single focal plane
#        * T: single timepoint
#
#        Result: Average intensity of given pixels for each channel in the range, at given focal plane and timepoint.
#
#
#        ***Example 2:***
#        * X, Y: range of pixel coordinates
#        * C: range of channels with `MAX` reduction
#        * Z: range of focal planes
#        * T: range of timepoints
#
#        Result: Maximum intensity across all given channels for each (Z,T) plane combination.
#      tags:
#        - Hyperspectral
#      parameters:
#        - $ref: '#/components/parameters/path-filepath'
#        - $ref: '#/components/parameters/header-etag'
#      requestBody:
#        required: true
#        content:
#          application/json:
#            schema:
#              oneOf:
#                - title: region-by-zoom-and-region
#                  allOf:
#                    - $ref: '#/components/schemas/input-region'
#                    - $ref: '#/components/schemas/footprint-plane-selection'
#                - title: region-by-level-and-region
#                  allOf:
#                    - $ref: '#/components/schemas/input-level-region'
#                    - $ref: '#/components/schemas/footprint-plane-selection'
#                - title: region-by-zoom-and-tile-index
#                  allOf:
#                    - $ref: '#/components/schemas/target-zoom-tile-index'
#                    - $ref: '#/components/schemas/footprint-plane-selection'
#                - title: region-by-level-and-tile-index
#                  allOf:
#                    - $ref: '#/components/schemas/target-level-tile-index'
#                    - $ref: '#/components/schemas/footprint-plane-selection'
#                - title: region-by-zoom-and-tile-coordinates
#                  allOf:
#                    - $ref: '#/components/schemas/target-zoom-tile-coordinates'
#                    - $ref: '#/components/schemas/footprint-plane-selection'
#                - title: region-by-level-and-tile-coordinates
#                  allOf:
#                    - $ref: '#/components/schemas/target-level-tile-coordinates'
#                    - $ref: '#/components/schemas/footprint-plane-selection'
#      responses:
#        200:
#          $ref: '#/components/responses/hyperspectral-footprint'
#        304:
#          $ref: '#/components/responses/not-modified'
#        400:
#          $ref: '#/components/responses/problem-bad-request'
#        404:
#          $ref: '#/components/responses/problem-not-found'
#  /image/{filepath}/spectral/window:
#    post:
#      operationId: pims.api.spectral.show_spectral_window
#      summary: Get spectral window
#      description: |
#        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
#        with body content must be used.**
#
#        Get spectral window. The input spatial region is given in the same way as regular windows. In output, it is
#        returned the spectrum of each pixel belonging to the given spatial region.
#      tags:
#        - Hyperspectral
#      parameters:
#        - $ref: '#/components/parameters/path-filepath'
#        - $ref: '#/components/parameters/header-etag'
#      requestBody:
#        required: true
#        content:
#          application/json:
#            schema:
#              oneOf:
#                - title: region-by-zoom-and-region
#                  allOf:
#                    - $ref: '#/components/schemas/input-region'
#                    - $ref: '#/components/schemas/spectral-window-plane-selection'
#                - title: region-by-level-and-region
#                  allOf:
#                    - $ref: '#/components/schemas/input-level-region'
#                    - $ref: '#/components/schemas/spectral-window-plane-selection'
#                - title: region-by-zoom-and-tile-index
#                  allOf:
#                    - $ref: '#/components/schemas/target-zoom-tile-index'
#                    - $ref: '#/components/schemas/spectral-window-plane-selection'
#                - title: region-by-level-and-tile-index
#                  allOf:
#                    - $ref: '#/components/schemas/target-level-tile-index'
#                    - $ref: '#/components/schemas/spectral-window-plane-selection'
#                - title: region-by-zoom-and-tile-coordinates
#                  allOf:
#                    - $ref: '#/components/schemas/target-zoom-tile-coordinates'
#                    - $ref: '#/components/schemas/spectral-window-plane-selection'
#                - title: region-by-level-and-tile-coordinates
#                  allOf:
#                    - $ref: '#/components/schemas/target-level-tile-coordinates'
#                    - $ref: '#/components/schemas/spectral-window-plane-selection'
#      responses:
#        200:
#          $ref: '#/components/responses/hyperspectral-window'
#        304:
#          $ref: '#/components/responses/not-modified'
#        400:
#          $ref: '#/components/responses/problem-bad-request'
#        404:
#          $ref: '#/components/responses/problem-not-found'
...
components:
  parameters:
    header-etag:
      name: ETag
      in: header
      description: |
        The RFC 7232 ETag header field in a response provides the entity-tag of
        a selected resource. The entity-tag is an opaque identifier for versions
        and representations of the same resource over time, regardless whether
        multiple versions are valid at the same time. An entity-tag consists of
        an opaque quoted string, possibly prefixed by a weakness indicator (see
        [RFC 7232 Section 2.3](https://tools.ietf.org/html/rfc7232#section-2.3)).
      required: false
      schema:
        type: string
    header-x-image-size-safety:
      name: X-Image-Size-Safety
      in: header
      description: |
        This header provides hints about the way the server has to deal with too large image responses.
        * `SAFE_REJECT` - Reject too large image response and throw a `400 Bad Request`.
        * `SAFE_RESIZE` - Resize the image response to an acceptable image size. Information about the resize are
        returned in `X-Image-Size-Limit` header.
        * `UNSAFE` - **At your own risk!** Try to fulfill the request but can cause unintended side effects
        (unreadable response, server slown down, server failure, ...). It should only be used in rare controlled
        situations.
      required: false
      schema:
        default: SAFE_RESIZE
        type: string
        enum:
          - SAFE_REJECT
          - SAFE_RESIZE
          - UNSAFE
    header-x-annotation-origin:
      name: X-Annotation-Origin
      in: header
      description: |
        This header give the origin coordinate system in which are described provided annotations.
      required: false
      schema:
        default: LEFT_TOP
        type: string
        enum:
          - LEFT_TOP
          - LEFT_BOTTOM
    path-filepath:
      name: filepath
      in: path
      description: The file path, relative to server base path.
      required: true
      schema:
        type: string
        format: path
      example: '123/my-file.ext'
...
  schemas:
...
    spectral-window-plane-selection:
      type: object
      properties:
        channels:
          description: |
            Image channels used to render the response. This parameter is interpreted as a set such that duplicates
            are ignored.

            **If this parameter represents a set of size > 1, `z_slices` and `timepoints` have to be single indexes.**

            **By default,** all channels are considered.
          allOf:
            - $ref: '#/components/schemas/multiple-channel-indexes'
        z_slices:
          description: |
            Image focal planes used to render the response. This parameter is interpreted as a set such that duplicates
            are ignored.

            **If this parameter represents a set of size > 1, `channels` and `timepoints` have to be single indexes.**

            **By default,** the median focal plane considered.
          allOf:
            - $ref: '#/components/schemas/multiple-z-slice-indexes'
        timepoints:
          description: |
            Image timepoints used to render the response. This parameter is interpreted as a set such that duplicates
            are ignored.

            **If this parameter represents a set of size > 1, `z_slices` and `channels` have to be single indexes.**

            **By default,** the first timepoint considered.
          allOf:
            - $ref: '#/components/schemas/multiple-timepoint-indexes'
...
    target-length-object-required:
      type: object
      required:
        - length
      properties:
        length:
          $ref: '#/components/schemas/target-length'
...
    annotations-object-basis:
      type: object
      required:
        - annotations
      properties:
        annotations:
          $ref: '#/components/schemas/annotations'
...
    mask-by-length:
      allOf:
        - $ref: '#/components/schemas/annotation-mask-basis'
        - $ref: '#/components/schemas/target-length-object-required'
    mask-by-width:
      allOf:
        - $ref: '#/components/schemas/annotation-mask-basis'
        - $ref: '#/components/schemas/target-width-object-required'
    mask-by-height:
      allOf:
        - $ref: '#/components/schemas/annotation-mask-basis'
        - $ref: '#/components/schemas/target-height-object-required'
    mask-by-zoom:
      allOf:
        - $ref: '#/components/schemas/annotation-mask-basis'
        - $ref: '#/components/schemas/target-zoom-object-required'
    mask-by-level:
      allOf:
        - $ref: '#/components/schemas/annotation-mask-basis'
        - $ref: '#/components/schemas/target-level-object-required'
...
    crop-by-length:
      allOf:
        - $ref: '#/components/schemas/mask-by-length'
        - $ref: '#/components/schemas/annotation-crop-basis'
    crop-by-width:
      allOf:
        - $ref: '#/components/schemas/mask-by-width'
        - $ref: '#/components/schemas/annotation-crop-basis'
    crop-by-height:
      allOf:
        - $ref: '#/components/schemas/mask-by-height'
        - $ref: '#/components/schemas/annotation-crop-basis'
    crop-by-zoom:
      allOf:
        - $ref: '#/components/schemas/mask-by-zoom'
        - $ref: '#/components/schemas/annotation-crop-basis'
    crop-by-level:
      allOf:
        - $ref: '#/components/schemas/mask-by-level'
        - $ref: '#/components/schemas/annotation-crop-basis'
...
    drawing-by-length:
      allOf:
        - $ref: '#/components/schemas/mask-by-length'
        - $ref: '#/components/schemas/annotation-drawing-basis'
    drawing-by-width:
      allOf:
        - $ref: '#/components/schemas/mask-by-width'
        - $ref: '#/components/schemas/annotation-drawing-basis'
    drawing-by-height:
      allOf:
        - $ref: '#/components/schemas/mask-by-height'
        - $ref: '#/components/schemas/annotation-drawing-basis'
    drawing-by-zoom:
      allOf:
        - $ref: '#/components/schemas/mask-by-zoom'
        - $ref: '#/components/schemas/annotation-drawing-basis'
    drawing-by-level:
      allOf:
        - $ref: '#/components/schemas/mask-by-level'
        - $ref: '#/components/schemas/annotation-drawing-basis'
          
    # Files, metadata
...
  responses:
    problem-bad-request:
      description: Bad request - Input payload fails business logic validation or generic error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/invalid-parameters-problem'
    problem-not-found:
      description: Not Found - The resource is not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/problem'
    problem-not-acceptable:
      description: |
        Not Acceptable - The resource can only generate content not acceptable according to the Accept
        headers sent in the request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/problem'
...
    not-modified:
      description: Not Modified
...
    annotation-crop:
      description: OK - The annotation crop
      headers:
        If-None-Match:
          $ref: '#/components/headers/If-None-Match'
        Cache-Control:
          $ref: '#/components/headers/Cache-Control'
        Vary:
          $ref: '#/components/headers/Vary'
        X-Image-Size-Limit:
          $ref: '#/components/headers/X-Image-Size-Limit'
      content:
        image/jpeg:
          schema:
            $ref: '#/components/schemas/generic-image'
        image/png:
          schema:
            $ref: '#/components/schemas/generic-image'
        image/webp:
          schema:
            $ref: '#/components/schemas/generic-image'
...